refactor(#218): split logic and view in Canvas Page
