+++ apps/frontend/src/pages/CanvasPage.tsx
@@ -1,4 +1,3 @@
-import { useState, useRef } from 'react';
 import { observer } from 'mobx-react-lite';
 import {
   ReactFlow,
@@ -6,24 +5,15 @@ import {
   Background,
   BackgroundVariant,
   ConnectionMode,
-  useReactFlow,
 } from '@xyflow/react';
 import '@xyflow/react/dist/style.css';
 import {
-  useRelationships,
-  useTables,
-  useViewport,
-  useCanvasInitialization,
-  useCanvasKeyboard,
-  useCanvasNodes,
   TableNode,
   RelationshipMarker,
   Toolbar,
   RelationshipEditor,
   CustomControls,
   TablePreview,
-  type RelationshipConfig,
-  type Point,
   CustomSmoothStepEdge,
   CustomConnectionLine,
   FloatingButtons,
@@ -32,10 +22,9 @@ import {
   Memo,
   TempMemoPreview,
   MemoProvider,
-  useMemoContext,
+  useCanvasController,
 } from '@/features/drawing';
 import { ChatOverlay, ChatInput } from '@/components/Collaboration';
-import { collaborationStore } from '@/store/collaboration.store';
 
 const NODE_TYPES = {
   table: TableNode,
@@ -46,128 +35,38 @@ const EDGE_TYPES = {
   customSmoothStep: CustomSmoothStepEdge,
 };
 
-const CURSOR_THROTTLE_MS = 100;
-
-const CanvasContent = () => {
-  const { screenToFlowPosition } = useReactFlow();
-  const lastCursorSendTime = useRef<number>(0);
-
-  const [relationshipConfig, setRelationshipConfig] =
-    useState<RelationshipConfig>({
-      type: 'one-to-many',
-      isNonIdentifying: false,
-    });
-  const [activeTool, setActiveTool] = useState('pointer');
-  const [mousePosition, setMousePosition] = useState<Point | null>(null);
-  const [tempMemoPosition, setTempMemoPosition] = useState<{
-    flow: Point;
-    screen: Point;
-  } | null>(null);
-  const [chatInputPosition, setChatInputPosition] = useState<Point | null>(
-    null,
-  );
-
-  useCanvasInitialization();
-
-  useCanvasKeyboard({
-    chatInputPosition,
-    mousePosition,
-    activeTool,
-    setChatInputPosition,
-  });
-
-  const { handleMoveEnd } = useViewport();
-  const { tables, addTable, onTablesChange } = useTables();
-  const { memos, onMemosChange, createMemo } = useMemoContext();
-
+const CanvasContent = observer(() => {
   const {
-    relationships,
-    selectedRelationship,
-    onConnect,
-    onRelationshipsChange,
-    onRelationshipClick,
-    onReconnectStart,
-    onReconnect,
-    onReconnectEnd,
-    updateRelationshipConfig,
-    deleteRelationship,
-    changeRelationshipName,
-    setSelectedRelationship,
-  } = useRelationships(relationshipConfig);
-
-  const { nodes, handleNodesChange } = useCanvasNodes({
-    tables,
-    memos,
-    onTablesChange,
-    onMemosChange,
-  });
-
-  const handleMemoCancel = () => {
-    setTempMemoPosition(null);
-  };
-
-  const handleMemoCreate = (content: string) => {
-    if (tempMemoPosition) {
-      createMemo(tempMemoPosition.flow, content.trim());
-      setTempMemoPosition(null);
-    }
-  };
-
-  const handleChatSend = (message: string) => {
-    collaborationStore.sendMessage(message);
-    setChatInputPosition(null);
-  };
-
-  const handleChatCancel = () => {
-    setChatInputPosition(null);
-  };
-
-  const handlePaneClick = (e: React.MouseEvent) => {
-    if (tempMemoPosition) {
-      handleMemoCancel();
-      return;
-    }
-
-    if (activeTool !== 'table' && activeTool !== 'memo') return;
-
-    const flowPosition = screenToFlowPosition({
-      x: e.clientX,
-      y: e.clientY,
-    });
-
-    if (activeTool === 'table') {
-      addTable(flowPosition);
-      setActiveTool('pointer');
-      setMousePosition(null);
-    } else if (activeTool === 'memo') {
-      const target = e.currentTarget as HTMLElement;
-      const rect = target.getBoundingClientRect();
-      setTempMemoPosition({
-        flow: flowPosition,
-        screen: {
-          x: e.clientX - rect.left,
-          y: e.clientY - rect.top,
-        },
-      });
-      setActiveTool('pointer');
-      setMousePosition(null);
-    }
-  };
-
-  const handleMouseMove = (e: React.MouseEvent) => {
-    const position = { x: e.clientX, y: e.clientY };
-    setMousePosition(position);
-
-    const now = Date.now();
-    if (now - lastCursorSendTime.current >= CURSOR_THROTTLE_MS) {
-      lastCursorSendTime.current = now;
-      const flowPosition = screenToFlowPosition({
-        x: position.x,
-        y: position.y,
-      });
-      collaborationStore.sendCursor(flowPosition.x, flowPosition.y);
-    }
-  };
+    state: {
+      relationshipConfig,
+      activeTool,
+      mousePosition,
+      tempMemoPosition,
+      chatInputPosition,
+      selectedRelationship,
+    },
+    setter: { setRelationshipConfig, setActiveTool, setSelectedRelationship },
+    data: { nodes, relationships },
+    handlers: {
+      handleNodesChange,
+      onRelationshipsChange,
+      handleMoveEnd,
+      onConnect,
+      onRelationshipClick,
+      onReconnect,
+      onReconnectStart,
+      onReconnectEnd,
+      updateRelationshipConfig,
+      changeRelationshipName,
+      deleteRelationship,
+      handleMemoCancel,
+      handleMemoCreate,
+      handleChatSend,
+      handleChatCancel,
+      handlePaneClick,
+      handleMouseMove,
+    },
+  } = useCanvasController();
 
   return (
     <>
@@ -276,7 +175,7 @@ const CanvasContent = () => {
       <ChatOverlay />
     </>
   );
-};
+});
 
 const CanvasPageComponent = () => {
   return (
